Author class
Added Author class to separated file
