+++ src/scripts/Post.ts
@@ -1,4 +1,5 @@
 import Utilities from './Utilities'
+import { Author } from './Author'
 
 class Post {
   public slug: string
@@ -21,25 +22,8 @@ class Post {
   }
 }
 
-class Author {
-  constructor(
-    public name: string,
-    public description: string,
-    public link: string,
-    public image: string
-  ) {
-    this.name = name;
-    this.description = description;
-    this.link = link;
-    this.image = image;
-  }
 
-  getAuthorName() {
-    return this.name
-  }
-}
 
 export {
-  Post, 
-  Author
+  Post
 }
