+++ src/services/NetworkService.ts
@@ -1,6 +1,6 @@
 import RequestService from './RequestService'
 
-const API = 'https://api.rss2json.com/v1/api.json?rss_url=https://medium.com/feed/@freecodecamp';
+const API = 'https://api.rss2json.com/v1/api.json?rss_url=https://medium.freecodecamp.org/feed';
 
 class NetworkService {
