+++ src/templates/author.ts
@@ -1,11 +1,11 @@
-import { Author } from '../scripts/Post'
+import { Author } from '../scripts/Author'
 
 export default (author: Author) => {
   return `
     <div class="author">
       <figure class="author__avatar">
         <a href="${author.link}" target="_blank" title="${author.name}">
-          <img src="${author.image}" width="64" height="64" alt="${author.name}"/>
+          <img src="${author.image}" width="64" alt="${author.name}"/>
         </a>
       </figure>
       <div class="author-meta">
