+++ src/views/index.html
@@ -36,58 +36,47 @@
       body {
         font-family: 'Playfair Display', serif;
         font-size: 1em;
-      }
-
-      .page {
-          width: 100%;
-          min-height: 100vh;
-          float: left;
-          overflow-x: hidden;
+        overflow-x: hidden;
       }
 
       a {
-          text-decoration: none;
-          color: #000;
-      }
-
-      #medium {
-          width: 100%;
-          float: left;
+        text-decoration: none;
+        color: #000;
       }
 
       .post {
-          width: 100%;
-          height: 100vh;
-          max-width: 50%;
-          background-color: white;
-          padding: 3% 5%;
-          display: flex;
-          flex-direction: column;
-          justify-content: space-between;
+        width: 100%;
+        height: 100vh;
+        max-width: 50%;
+        background-color: white;
+        padding: 3% 5%;
+        display: flex;
+        flex-direction: column;
+        justify-content: space-between;
       }        
 
       .post:not(:last-child) {
-          border-right: 2px solid lightgray;
+        border-right: 2px solid lightgray;
       }
 
       .post:hover {
-          background-color: lightcyan;
+        background-color: lightcyan;
       }
 
       .post__date {
-          display: block;
-          margin: 0 0 20px;
+        display: block;
+        margin: 0 0 20px;
       }
 
       .post__title {            
-          font-size: 2.5vw;
-          font-weight: 700;
-          margin: 0 0 40px;
-          max-width: 90%;
+        font-size: 2.5vw;
+        font-weight: 700;
+        margin: 0 0 40px;
+        max-width: 90%;
       }
 
       .post__author {
-          display: block;
+        display: block;
       }
 
       .author {
